feat(webvitals): Update normalize_performance_score to accept components with 0 weight and automatically normalize weights to a sum of 1 (#2756)
- Updates `normalize_performance_score` to accept a
`PerformanceScoreProfile` with components of 0 weight. This is needed
because we must store 0 weight for unused component to ensure that the
sum of all `avg(measurements.score.weight.*)` equals `1` at any given
time. Not storing 0 allows the possibility that the total sum of avg
weights exceeds 1.
- Updates `normalize_performance_score` to accept a
`PerformanceScoreProfile` with a component weights sum of any number
greater than 0. Previously, we expected the component weights from the
`PerformanceScoreProfile` config to always sum up to 1. This change
automatically normalizes the stored component weights so that the sum of
stored weights will always equal 1.