feat(server): Return global config status for downstream requests (#2765)
follow-up to https://github.com/getsentry/relay/pull/2697


we now have a mechanism to delay envelope processing until a global
config is available, however, we currently send back a default global
config if one is requested from downstream, which kind of defeats the
purpose if the downstream relay believes it has a valid global config
but it doesn't.
 
To fix this, this pr will return an additional flag whether the config
it sends is ready or not, up to date downstream relays will not use them
but instead keep trying to get a ready global config and not process
envelopes until they do. Older relays won't deserialize the status flag
and their behaviour will be as before.