feat(normalization): Introduce event normalization entry point (#2774)
This PR introduces an entry point for event normalization. It currently
only runs what previously was "light normalization", but it should
contain all normalization logic in the future. This PR doesn't contain
any functional changes.

Previous PRs approached the goal by introducing a new processor and
calling it directly. This not only makes a more difficult-to-use API
(`process_value` with a processor, a processing state, and the processor
itself), but also allows running normalization with non-top-level items;
for example, it'd be possible to run normalization on a standalone span.
Normalization must be controlled from top-level items and a processor
doesn't provide such control a regular function does. Internally, this
function uses other processors depending on what the goals are.