debug(spool): Add metrics and report an error (#2828)
Let's track the timings how long the envelope took to unspool. 
Also add error if the envelope was spooled longer than 2 hours.

Note: This can be chatty in case of the long lasting incident, but this
way we will be able to see if any of the envelopes are stuck in the
spool, and cannot be unspooled anymore.
  
This is temporary and just to help with debugging of the on-disk spool.