build: Update axum and http (#2844)
We would like to update the Sentry SDK, which requires to update `axum`
and related dependencies. As `reqwest` has not updated to the latest
version of `http` and `hyper` yet, we will temporarily run two versions
of the frameworks side-by-side and convert between their types in the
forward endpoint.

The `axum-server` library also has not released their latest version
yet. Since the only thing needed from axum-server is graceful shutdown,
we implement this ourselves now following the example in the `axum`
repository.