feat(processor): Add a metric to track the time of all messages (#2877)
In order to estimate the throughput and capacity of the envelope
processor service, we're tracking the number and duration of all
processed messages. This should allow us to assess when the CPU pool is
reaching its limit.

The metric is tracked as `processor.message.duration` and has a
`message` tag containing the name of the variant / struct that is being
processed. This allows to break down the volume and duration of each of
the messages by type.