ref(server): Separate methods for encoding metrics (#2890)
This introduces two methods for flushing metric buckets:

1. The version for external and PoP Relays that performs partitioning,
   batching, and sends envelopes.
2. The version for processing Relays that performs cardinality limiting,
   rate limiting, and sends kafka messages directly. Batching occurs in
   the store forwarder.

To submit metric buckets to kafka more efficiently, there's a new
message `StoreMetrics` that can take in-memory buckets. It splits
buckets internally so that the caller doesn't have to take care of that
anymore. This is a fire-and-forget message that does not require the
caller to await the result.

The `StoreForwarder` therefore is now responsible to log outcomes for
metric buckets that cannot be produced.