ref(metrics): Change flush into a cross-project message (#2894)
The metrics aggregator service now flushes the buckets of all projects
in a single `FlushBuckets` message. Previously, this message was emitted
separately for every project.

This change allows the processor to batch buckets from multiple projects
into a single request. This change is not yet included in this PR and
will be added in a follow-up.