feat(server): Send metrics via the global endpoint (#2902)
Introduces an option `http.global_metrics`. When enabled and not in
processing mode, Relay sends metrics to the global batch endpoint at
`/api/0/relays/metrics/` instead of envelopes. This endpoint allows for
batched submission of metrics from multiple projects, which should
reduce the overall number of requests.

# Bug Fixes

This change contains additional bug fixes that were discoverd during
implementation:

- The batch metrics endpoint no longer requires the `emit_outcomes`
  outcomes flag to be set. This was invalid copy & paste from the
  outcomes endpoint.
- Requests with HTTP encoding received a signature computed from the
  compressed body. However, Relay requires the signature on the
  uncompressed body.

# Details

Building the request occurs in the `EnvelopeProcessor` in place of
building envelopes in the following steps:

1.  Iterate buckets from all projects and stream them into partitions.
2.  While collecting, check the size of the partition. As soon as a
    partition reaches the batch size limit, flush the partition eagerly.
    Buckets at the border may be split.
3.  At the end flush all remaining partitions with data.
4.  To flush a partition, JSON-encode the list of buckets and optionally
    apply HTTP encoding (compression).
5.  Submit the `SendMetricsRequest` with the payload and outcome
    metadata directly to the upstream.
6.  In the response callback, log outcomes directly. Therefore, the
    request does not have to be awaited.

In processing mode, Relay still produces to Kafka. The configuration
option does not have an effect in processing mode.

A note on stability: This endpoint and functionality is meant for
operation at scale within a distributed Sentry installation. At this
moment, it is not recommended to enable this option for external Relays.