ref(spans): Reduce number of tags on span metrics (#2907)
During the development of Starfish we've repeatedly added new tags to
span metrics without restricting the conditions under which a tag is set
(that is, what queries actually require a tag).

This PR introduces a new config for the `exclusive_time` and
`exclusive_time_light` metrics, which limits the number of cases in
which the metrics and / or their tags are extracted. Some tags are
removed entirely.

For now, the new config will only be enabled for projects that have the
`all-modules` feature flag enabled, i.e. sentry-internal. After we've
verified that the new config covers all required cases, we can enable
the new config for all orgs.

ref: https://github.com/getsentry/team-ingest/issues/252