feat(app-start): Add Android app start-related spans (#2927)
These spans are nested operations under the app.start.* spans on
Android necessary for debugging start time operations. This only
adds the application portion of the spans. We are currently awaiting
a new span that tracks the system portion of the app start and I
will add that as another span op afterwards.