feat(inbound-filters): Create new legacy browser filters (#2950)
this pr adds in new legacy browser filters. These new filters will be
more up to date than the existing ones, and would allow for easier
updates to the filters if needed. right now, the filters are defined as
both a browser and a version ("safari_pre_6"). now the filters will just
be for a browser, and we will have easier control over what versions to
filter- we no longer have to make a new filter, we can just update what
values get filtered. this can both be a short term and longer term fix
to the problem with legacy browsers. It will allow us to move over to
new filters in the short term, while supporting updates in the long term
if needed (since the generic filters project doesn't have a timeline
right now).


closes https://github.com/getsentry/sentry/issues/63053