feat(cardinality): Implement quota/limit system for cardinality limiter (#2972)
Implements a similar system we already have for rate limits for the
cardinality limiter. Instead of a single global limit now quotas can be
defined matching on different scopes (currently only org) and
namespaces.

This allows us to support multiple scopes in the future and right now to
support different limits per namespace.

Metrics are now tagged with the limit id instead of the namespace of the
metric being checked. Not sure if we want to change this in the future.