ref(cardinality): Track rejections in an accumulator (#2978)
Instead of returning an iterator of accumulated rejections track all
rejections in the accumulator directly. This allows the caller to pass
in whatever accumulator it needs and skips at least one round of
allocations (Before `Vec<Rejection> -> HashSet<usize>`, now it is just
`HashSet<usize>`. This is also a preparation for when we want to track
more than just which element was rejected, we can easily include the
`CardinalityLimit` information in the argument to the accumulator.