fix(spans): Apply max. SQL expression depth after simplification (#3003)
Repeating OR/AND conditions are collapsed, but if the maximum query
depth is reached, the limit is applied before we can collapse it. This
leads to scrubbed queries like:

```sql
WHERE ((.. OR (..) OR (.. = %s AND b = %s) OR (a = %s AND b = %s))
```

Fix by splitting the query visitor into two steps (first normalize, then
reduce depth).

This PR also removes redundant parentheses around binary operations
(e.g. `a OR (b OR c)`).

ref: [internal
issue](https://www.notion.so/sentry/Collapse-identical-OR-conditions-in-queries-40786f8cddb145e398e8ad2d5eb66522)

---------

Co-authored-by: Oleksandr <1931331+olksdr@users.noreply.github.com>