ref(cardinality): Do not lookup scoped cache for every update (#3043)
Previously every `accept` call to insert a hash into the cache would
lookup the scoped cache again. Using parking lot we can store an owned
reference to the `ScopedCache`.

We're already using `parking_lot` elsewhere in Relay (directly and
indirectly), it also get's rid of the lock poisoning and it is generally
more performant than the std locks.