feat(app-start): Tag spans with start type (#3027)
We want to tag all spans during an app start with the start type. This
is because on Android, the span ops don't reflect the start type but we
still want to filter spans by their type.

Start type can be derived from the transaction if there is a measurement
associated, or in the future we will add `app_start_type` to span data
and that should take precedence.