feat(processor): Add initial types into processing pipeline (#3078)
This PR puts types restrictions on some functions, which makes it harder
to get wrong. There is no functional changes over here. The code that
was moved around wasn't changed, it's just to make it easier to create
the current state in more control manner.

This is just initial setup, which should evolve once we move to
`TypedEnvelope` (will be described in
https://github.com/getsentry/team-ingest/issues/281)