fix(project-cache): Increase project cache worker threads (#3150)
The project cache Runtime only has a single thread. The idea is since it
only ever works on a single task at a time it only needs a single
thread.

But the project cache does spawn more tasks via `tokio::spawn` in 
`handle_request_update` (a substantial amount), this interferes with the single thread only
working on the project cache backlog.

This is a bit of a stab in the dark to fix processing delays in our
dedicated project config Relays.