test(global-config): Start roundtrip test from serialization (#3159)
This PR updates the global config roundtrip test order, deserializing
and serializing it again, and running the validation on serialized
values.

The main goal is to not derive `PartialEq` in `GlobalConfig` anymore. It
was only used in this test and comparing strings results in it being
unused. Additionally, it allows us to introduce new fields in
`GlobalConfig` that don't derive `PartialEq`.

Additional benefits:
- Remove helper function `mock_quota` and all associated imports.
- Make the test easier to update and extend, by updating a string
instead of typing it completely.
- Make the protocol easier to understand. By fully typing the global
config, custom (de)serialization may generate confusion on what the
protocol is, in case we decide to add it in the future.