fix(gocd): Fixing sentry checks by ensuring they use the right release (#3194)
This change stems from the realization that Relay doesn't simply use the
SHA when creating a release, but also the version. As a result, the
complexity of the scripts had to increase a bit to accommodate this.

Things worth mentioning:
* Since the Sentry project for PoPs is pop-relay, but the service is
referred to as relay-pop in google storage, we needed another env var to
represent the service rather than using the sentry-project's name.
* Due to there being three env vars that each contain multiple,
space-delimited inputs, I felt it was worthwhile to combine them into
tuples to mitigate confusion.
* To get the release name, I use the revision and service name to search
google storage for the corresponding release name. If there is an easier
way, please let me know.
* I realized that since Processing and PoP's pipelines are independent
there is no guarantee that they will be deployed at the same time or
will even be in sync (though they ideally should be). As a result, using
the other service's Sentry release as a signal could be problematic,
especially if the pipelines are not kept in sync. One step I took to
help alleviate this issue is not cause the check to fail if a given
release name is not found.
* Similar to the point directly above, assuming we use Sentry releases
as signals from both services, the longer the period between their
respective deploys, the more likely it is that a new error will occur
and thus cause the new error check to fail. The new error check was
originally designed to be run shortly after a deploy goes out since a
new error occurring in that short time period is often indicative of a
problematic change. As a result, the longer the release has been live
before we check, the less useful this signal becomes. While this is not
inherently problematic, it is something to consider when it comes to how
best we should use these pipelines.


#skip-changelog