fix(gocd): Moving comment to fix GoCD script parsing (#3211)
This PR addresses an issue with GoCD's handling of inline comments in
scripts. GoCD requires `#` characters to either denote a variable or be
prefixed with an additional `#` character. Because comments in scripts
are quite common, we have added a utility that automatically adds an
additional `#`, but it is quite naive and doesn't understand nested
comments. As a result, I decided to move the nested comment up to the
top for the time being to unblock progress.

#skip-changelog