ref(cardinality): Passive mode per cardinality limit (#3199)
Makes it possible to mark individual cardinality limits per project to
be marked as passive.

Initially the idea was not to add the `passive` flag to the individual
limit and only enforce the passive property/mode in the processor. But
it turns out this is more complicated than just directly discarding a
passive result in the cardinality limiter. This was made easy due to
already having the correct abstraction (`Rejections` trait) in place, it
just needed to be slightly updated to be passed the limit instead of
just the limit id. This approach also allows us to tag all metrics and
cardinality errors with a passive tag.