chore(mobile-score): Add test documenting mobile score profile
Adds a test documenting a mobile score profile we want to experiment
with. All fields will be optional for now until we figure out the
combinations of measurements that are sent together.

The next step will be to add a PR that tags these scores with a score
profile name that we can use for versioning while prototyping or making
changes to perf score profile weights.