feat(mobile-score): Add profile name as tag
When we have data to write for tags, also write the profile name. This
can be used in the future for versioning while we prototype different
profiles for scores.

This should have low cardinality because we would only ever expect a few
types of profile versions to be in the while at a time.