feat(metrics): Add metadata for buckets (#3254)
Adds `BucketMetadata` to the metrics `Bucket`. The metadata can be used
to track meta information about the bucket, like amount of aggregations
(total merges).

The metadata currently only tracks the amount of "merges".

Bucket metadata is a purely Relay internal detail, clients are not
expected to send it and Relay will ignore and reset the metadata if it
is not coming from a trusted/internal Relay. This makes it possible to
to implement the [received timestamp for
metrics](https://github.com/getsentry/team-ingest/issues/265) while at
the same time forwarding the metadata correctly from PoPs to Processing
Relays without any additional configurations.