ref: Kafka topic configuration
As part of the kafka control plane project, Kafka config for all Sentry services
should be automatically generated from a single source of truth in ops/sentry-kafka-schemas.

The system knows about the default topic name + override value per environment,
but is not aware of the Relay specific values like "events" and "metrics".

This is aligned with the configuration format of Sentry, Snuba and super-big-consumers.