feat(mobile-score): Add profile version as tag (#3249)
Add a `version` property to perf score profiles so we can filter and
separate changes to the profiles. The `version` gets assigned as
`sentry.score_profile_version` in a transaction's tags. This would be
useful if we wanted to hard cut over to a new version instead of
blending users in, or just for creating test bed profiles. In our case,
we want to send profiles with `version: mobile.alpha` so we can iterate
on profiles but avoid polluting the end user results.

This should have low cardinality because we would only ever expect a few
types of profile versions to be in the while at a time.

This adds a new string to the metrics indexer. The corresponding PR in
`sentry` is: https://github.com/getsentry/sentry/pull/66795