ref(normalization): Remove StoreProcessor (#3097)
This PR removes the `StoreProcessor`, bringing us one step closer to
removing the store normalizer entirely.

The biggest difference is that normalization of default attributes is
now in `event::normalize` and runs in all relays, not just processing.

### Design choices

- There's a `StoreConfig` struct in the c-abi. Setting the configuration
from librelay should happen with a single object and Relay uses 3 (2 for
validation and 1 for normalization), so we can't get rid of it in the
current setting.
- The normalization config has a `is_last_normalize`. Relay needs to
know whether it's the last normalization step to e.g. remove unknown
fields not to break future compatibility. This flag is currently enabled
in processing relays and librelay calls, and it's disabled by default.

### Future work
- Move `normalize::mod` contents to their appropriate place. Currently,
it contains a few types, a couple of functions used in other places, and
tests validating the functionality in other modules.
- Make running normalization configurable, and only run it in PoP
relays.