ref(metrics): Validate metric units (#3390)
Validates that metric units consist of ASCII word characters.
Previously, more lenient validation was applied. Since Relay is now the
only place to normalize MRIs, we're enforcing the rules more strictly.

Note that the spec for what's considered a valid unit has not changed.

In a future change, Relay will start to emit errors or meta data for
invalid MRIs and tags so that users receive some feedback and can fix
potential issues.