build(sentry): Vendor the sentry-tower integration (#3395)
Vendors the `sentry-tower` integration and upgrades the Sentry SDK to
`0.32.x` without upgrading axum. Going forward, this allows us to keep
the Sentry SDK up-to-date independently of the web framework.

The integration's code was copied from version `0.31.8`, which is the
last release matching our version of `axum`. Minor adjustments to the
imports were made.

This will be removed once `axum` is upgraded to latest.