ref(filters): Access fields through trait (#3397)
Currently, inbound filters only work for event types, that is error
events and transactions. In the near future we will want inbound filters
to apply to other data types as well, such as standalone spans and
session replays. To prepare for this generalization, make filters act on
a trait that supplies the necessary field instead of the `Event` itself:

```rust
// `Filterable` for legacy inbound filters,
// `Getter` for generic inbound filters.
pub fn should_filter<F: Filterable + Getter>(item: &F, ...); 
```

ref: #3317

---------

Co-authored-by: Iker Barriocanal <32816711+iker-barriocanal@users.noreply.github.com>