feat(perf): Extract scrubbed IP addresses into the `span.domain` tag (#3383)
tl;dr URLs with IP domains like `http://8.8.8.8/data` get the
`span.domain` tag of `*.*.*.*`. Also a small refactor to support this.

Right now, if a span's description looks like `GET http://8.8.8.8/data`
the scrubbed `span.description` is `GET http://*.*.8.8/` and
`span.domain` is `None`. This is inconsistent, and also makes for a bad
user experience because people cannot find requests to certain IPs by
the domain.

There are two causes:

1. IP addresses in descriptions are ran through domain scrubbing, even
though they're not domains. This turns `8.8.8.8` into `*.*.8.8` because
the scrubber thinks the last two 8s are a TLD
2. The scrubbed descriptions are parsed by `URL` and are invalid. The
parsing fails, and no `domain` tag is produced.

This PR makes some refactors, and turns off scrubbing of IP host
strings.

## Changes

- `normalize_domain` --> `scrub_host`. Instead of a function that takes
a domain string and a port and does the asterisking, the function
accepts a `Url::Host` and does different things based on whether it's an
IP or a qualified domain. This is nicer because the name is clearer and
the function does fewer things
- `scrub_domain_name` is a new function that _only_ replaces a string
with asterisks. This makes it possible to call this function
independently of other scrubbing logic (e.g., to forgo calling it for IP
addresses)
- `concatenate_host_and_port` is a simple concatenator, since I saw this
logic in a few places
- `scrub_ipv4` and `scrub_ipv6` that fully scrub out IP addresses

Co-authored-by: David Herberth <david.herberth@sentry.io>