fix(config): Move normalization config to top level (#3428)
Normalization config was inside the `processing` config, which requires
specific config for processing Relays. This is convenient as it raises
errors during Relay startup when core features are not properly
configured. However, normalization config lives inside the processing
one and it's not possible to enable it in non-processing relays. This PR
moves it to a separate top-level config to make that possible.