fix(runtime): Bump service runtimes with tokio::spawn from 1 to 2 threads (#3454)
Stats are showing a backlog in the router service and some health checks
timing out. The router service makes use of `tokio::spawn` it is
possible that the backlog is made worse by the runtime being single
threaded.

The backlog leads to health checks timing out.

This change did have very positive effects on the project cache. If it
doesn't seem to improve the situation we can revert it again.