fix(server): Check rate limits with namespace in fast-path (#3447)
Checks rate limits for metrics with their namespace in the fast-path.
Previously, the namespace was not passed to the check function, which
caused all namespace-specific rate limits to be skipped during the
check.

This also simplifies `check_buckets` and removes an indirection to an
inner method. The signature of the method could no longer be fulfilled,
as now there can be partial rate limiting inside of the function.
Originally, this function was introduced to test that no error log is
emitted for one of the failure branches. However, the test did not
actually assert that and so we remove this test and simplify instead.