feat(spans): Respect country code TLDs when scrubbing span tags (#3458)
Domains like `application.co.uk` get scrubbed out to `*.co.uk` which is
not useful! `.co.uk` is a ccTLD, and should be left intact. I'm using
the `psl` crate, which is a pretty popular one, and it powers some
popular but higher-level libraries like `addr`. The new scrubber knows
about ccTLDs and leaves them alone.

---------

Co-authored-by: Joris Bayer <joris.bayer@sentry.io>