feat(mobile-ui): Collect frame delay metric (#3472)
Adds a new metric for collecting the `frames.delay` property from a
span's `data` object in the Starfish Mobile UI module.

I've added the tags that I think are necessary to query and filter for
this metric, and they are common filters for other metrics in this
config. Also noted is that the data is in the `data` object, similar to
#3473 and thus I modified measurement extraction to copy this value over
to span measurements.