fix(spans): Extract http status code from span data when provided as int (#3491)
Some SDKs send this field as an integer, which gets converted to `None`
by the `v.as_str()` call.

I added code to catch if the value is an integer and convert it to a
string. In terms of testing, I tested the function which extracts the
status, but also added sample spans to the mobile event to generate a
snapshot validating both string or integers are extracted.

The issue was identified as a difference between Android and iOS events.