fix(spans): Normalize segment span ID (#3531)
`SpanId` is represented as a string internally, so if we construct it
without going through `Annotated::from_value`, lowercase span IDs and
uppercase span IDs will be treated as different. This is problematic in
`set_segment_attributes`, which compares the span's ID to its segment
ID.

We should probably represent `SpanId` as `[u8; 8]`, or at least go
through `Annotated::from_value` for all Otel types, but this PR should
do as a quick fix.