fix(metric-stats): Only emit volume metadata for the first bucket view (#3609)
Splitting a bucket into multiple via a `BucketView` brings the problem
that some of the metadata can't be attributed correctly anymore
(aggregation is a lossy process for metadata). To solve this problem for
the volume metadata, it will only be attached to the bucket view
containing the start of the bucket.