feat(spans): Extract additional user fields for spans (#3599)
Trying to bring the user fields on spans closer to transactions. User
alone is not ergonomic and we'd like to be able to search by user email,
username etc. Transactions also has the user's ip address but the spans
table currently does not have the ipv4 and ipv6 columns for it so going
to hold off on it as the other fields are more useful.