Send `attachment` data inline when possible (#3654)
For `invidual_attachments`, we can now send the `data` inline if it is
small enough. This avoids having to store temporary `attachment_chunks`
in the processing redis store.

Right now, every "event" attachment is sent with outlined chunks, even
for things like logfiles, even though only minidumps and other
processing-relevant attachments need to be treated that way.
It might be possible to also inline logfiles with more logic. That way,
such attachments could be saved as `EventAttachment` directly without
putting load on the attachments store. Doing that however would create
problems when such events are discarded within the rest of the pipeline.

Support for consuming the `data` property is being added in
https://github.com/getsentry/sentry/pull/71522 to the Sentry codebase.